peer-selection: moved peerChurnGovernor to its own module
